requirements.txt
pytest==8.2.0
allure-pytest==2.13.5
pytz==2024.1
pillow==11.1.0
requests==2.32.3
flask==3.1.0
flask-jwt-extended==4.7.1
sqlalchemy==2.0.30